ZK-93: Moved handleCreateEntryTagWithValue into EntryTagPicker
